app/requirements.txt
sqlalchemy ~= 1.4.36
paho-mqtt ~= 1.6.1
